Move sender related tests to sender_test.go
