+++ exporter/sumologicexporter/exporter_test.go
@@ -82,7 +82,7 @@ func TestGetMetadata(t *testing.T) {
 	attributes.InsertString("additional_key2", "value2")
 	attributes.InsertString("additional_key3", "value3")
 
-	regexes :=  []string{"^key[12]", "^key3"}
+	regexes := []string{"^key[12]", "^key3"}
 	f, err := newFiltering(regexes)
 	assert.NoError(t, err)
 
@@ -125,169 +125,4 @@ func exampleLog() []pdata.LogRecord {
 	buffer[0].Body().SetStringVal("Example log")
 
 	return buffer
-}
-
-func TestSend(t *testing.T) {
-	test := getExporter(t, []func(req *http.Request){func(req *http.Request) {
-		body := extractBody(t, req)
-		assert.Equal(t, body, "Example log\nAnother example log")
-		assert.Equal(t, req.Header.Get("X-Sumo-Fields"), "test_metadata")
-		assert.Equal(t, req.Header.Get("X-Sumo-Client"), "otelcol")
-		assert.Equal(t, req.Header.Get("Content-Type"), "application/x-www-form-urlencoded")
-	}})
-	defer func() { test.srv.Close() }()
-
-	buffer := make([]pdata.LogRecord, 2)
-	buffer[0] = pdata.NewLogRecord()
-	buffer[0].InitEmpty()
-	buffer[0].Body().SetStringVal("Example log")
-	buffer[1] = pdata.NewLogRecord()
-	buffer[1].InitEmpty()
-	buffer[1].Body().SetStringVal("Another example log")
-	test.s.buffer = buffer
-
-	err := test.s.sendLogs("test_metadata")
-	assert.NoError(t, err)
-}
-
-func TestSendSplit(t *testing.T) {
-	test := getExporter(t, []func(req *http.Request){
-		func(req *http.Request) {
-			body := extractBody(t, req)
-			assert.Equal(t, body, "Example log")
-		},
-		func(req *http.Request) {
-			body := extractBody(t, req)
-			assert.Equal(t, body, "Another example log")
-		},
-	})
-	defer func() { test.srv.Close() }()
-	test.se.config.MaxRequestBodySize = 10
-
-	buffer := make([]pdata.LogRecord, 2)
-	buffer[0] = pdata.NewLogRecord()
-	buffer[0].InitEmpty()
-	buffer[0].Body().SetStringVal("Example log")
-	buffer[1] = pdata.NewLogRecord()
-	buffer[1].InitEmpty()
-	buffer[1].Body().SetStringVal("Another example log")
-	test.s.buffer = buffer
-
-	err := test.s.sendLogs("test_metadata")
-	assert.NoError(t, err)
-}
-
-func TestSendJson(t *testing.T) {
-	test := getExporter(t, []func(req *http.Request){func(req *http.Request) {
-		body := extractBody(t, req)
-		expected := `{"key1":"value1","key2":"value2","log":"Example log"}
-{"key1":"value1","key2":"value2","log":"Another example log"}`
-		assert.Equal(t, body, expected)
-		assert.Equal(t, req.Header.Get("X-Sumo-Fields"), "test_metadata")
-		assert.Equal(t, req.Header.Get("X-Sumo-Client"), "otelcol")
-		assert.Equal(t, req.Header.Get("Content-Type"), "application/x-www-form-urlencoded")
-	}})
-	defer func() { test.srv.Close() }()
-	test.se.config.LogFormat = JSONFormat
-
-	buffer := make([]pdata.LogRecord, 2)
-	buffer[0] = pdata.NewLogRecord()
-	buffer[0].InitEmpty()
-	buffer[0].Body().SetStringVal("Example log")
-	buffer[0].Attributes().InsertString("key1", "value1")
-	buffer[0].Attributes().InsertString("key2", "value2")
-	buffer[1] = pdata.NewLogRecord()
-	buffer[1].InitEmpty()
-	buffer[1].Body().SetStringVal("Another example log")
-	buffer[1].Attributes().InsertString("key1", "value1")
-	buffer[1].Attributes().InsertString("key2", "value2")
-	test.s.buffer = buffer
-
-	err := test.s.sendLogs("test_metadata")
-	assert.NoError(t, err)
-}
-
-func TestSendJsonSplit(t *testing.T) {
-	test := getExporter(t, []func(req *http.Request){
-		func(req *http.Request) {
-			body := extractBody(t, req)
-			assert.Equal(t, body, `{"key1":"value1","key2":"value2","log":"Example log"}`)
-		},
-		func(req *http.Request) {
-			body := extractBody(t, req)
-			assert.Equal(t, body, `{"key1":"value1","key2":"value2","log":"Another example log"}`)
-		},
-	})
-	defer func() { test.srv.Close() }()
-	test.se.config.LogFormat = JSONFormat
-	test.se.config.MaxRequestBodySize = 10
-
-	buffer := make([]pdata.LogRecord, 2)
-	buffer[0] = pdata.NewLogRecord()
-	buffer[0].InitEmpty()
-	buffer[0].Body().SetStringVal("Example log")
-	buffer[0].Attributes().InsertString("key1", "value1")
-	buffer[0].Attributes().InsertString("key2", "value2")
-	buffer[1] = pdata.NewLogRecord()
-	buffer[1].InitEmpty()
-	buffer[1].Body().SetStringVal("Another example log")
-	buffer[1].Attributes().InsertString("key1", "value1")
-	buffer[1].Attributes().InsertString("key2", "value2")
-	test.s.buffer = buffer
-
-	err := test.s.sendLogs("test_metadata")
-	assert.Nil(t, err)
-}
-
-func TestSendUnexpectedFormat(t *testing.T) {
-	test := getExporter(t, []func(req *http.Request){func(req *http.Request) {}})
-	defer func() { test.srv.Close() }()
-	test.se.config.LogFormat = "dummy"
-
-	buffer := make([]pdata.LogRecord, 1)
-	buffer[0] = pdata.NewLogRecord()
-	buffer[0].InitEmpty()
-	buffer[0].Body().SetStringVal("Example log")
-
-	err := test.s.sendLogs("test_metadata")
-	assert.Error(t, err)
-}
-
-func TestOverrideSourceName(t *testing.T) {
-	test := getExporter(t, []func(req *http.Request){func(req *http.Request) {
-		assert.Equal(t, req.Header.Get("X-Sumo-Name"), "Test source name")
-	}})
-	defer func() { test.srv.Close() }()
-
-	test.se.config.SourceName = "Test source name"
-	test.s.buffer = exampleLog()
-
-	err := test.s.sendLogs("test_metadata")
-	assert.Nil(t, err)
-}
-
-func TestOverrideSourceCategory(t *testing.T) {
-	test := getExporter(t, []func(req *http.Request){func(req *http.Request) {
-		assert.Equal(t, req.Header.Get("X-Sumo-Category"), "Test source category")
-	}})
-	defer func() { test.srv.Close() }()
-
-	test.se.config.SourceCategory = "Test source category"
-	test.s.buffer = exampleLog()
-
-	err := test.s.sendLogs("test_metadata")
-	assert.Nil(t, err)
-}
-
-func TestOverrideSourceHost(t *testing.T) {
-	test := getExporter(t, []func(req *http.Request){func(req *http.Request) {
-		assert.Equal(t, req.Header.Get("X-Sumo-Host"), "Test source host")
-	}})
-	defer func() { test.srv.Close() }()
-
-	test.se.config.SourceHost = "Test source host"
-	test.s.buffer = exampleLog()
-
-	err := test.s.sendLogs("test_metadata")
-	assert.Nil(t, err)
 }
